tgparser/requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
attrs==22.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
build==0.10.0
CacheControl==0.12.11
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cryptography==39.0.2
distlib==0.3.6
Django==4.2.7
dulwich==0.21.3
exceptiongroup==1.1.3
filelock==3.9.0
flower==1.1.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.1
httpx==0.25.1
humanize==4.8.0
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
kombu==5.3.2
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.0
pexpect==4.8.0
Pillow==10.1.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-telegram-bot==20.6
pytz==2023.3.post1
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.6
tornado==6.3.3
trove-classifiers==2023.3.9
typing_extensions==4.8.0
urllib3==1.26.15
vine==5.1.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
wcwidth==0.2.9
webencodings==0.5.1
zipp==3.15.0
